# make recurse_jsonify() work directly on types