--- a/chia/util/misc.py
+++ b/chia/util/misc.py
@@ -1,4 +1,3 @@
-import dataclasses
 from typing import Any, Dict, Sequence, Union
 
 from chia.util.streamable import recurse_jsonify
@@ -81,4 +80,5 @@ def get_list_or_len(list_in: Sequence[object], length: bool) -> Union[int, Seque
 
 
 def dataclass_to_json_dict(instance: Any) -> Dict[str, Any]:
-    return recurse_jsonify(dataclasses.asdict(instance))
+    ret: Dict[str, Any] = recurse_jsonify(instance)
+    return ret