circumventing the dataclasses.asdict() step. This enables simpler integration of non dataclasses into the Streamable and JSON protocols. Specifically, it opens the door for allowing native rust types to participate in `Streamable` (which includes json conversion).

As a bonus, this change avoids creating an intermediate dictionary, before converting it into JSON form.

A dictionary in JSON form is one that only contains `Dict`, `List`, `str` and `int`. `bytes32` objects are turned into hexadecimal strings.

This patch has a counterpart in `chia_rs`: https://github.com/Chia-Network/chia_rs/pull/18

**EDIT:**

I added a benchmark to ensure this doesn't make it more expensive to generate JSON messages. The benchmark calls `to_json_dict()` on `FullNode` objects of varying kinds (32768 combinations) timing those calls.

| before | after | after/before |
| --- | --- | --- |
| 336.15s | 101.90s | 0.303 |

So, a material speed-up