@xdustinface 2022-05-17
@arvidn Not sure if we need another benchmark here. Have you seen `benchmarks/streamable.py`? 

You can generate a record on `main` with `python streamable.py -o main.json` optionally add `-m to_json` and then compare it like below

```
python streamable.py -c main.json -m to_json

benchmarks: to_json, data: BenchmarkClass runs: 100, ms/run: 50, commit_hash: eea7e0764-dirty
mode       | µs/iteration | stdev µs/iteration % | avg iterations/run | stdev iterations/run %
to_json    | 294.22       |                 5.34 |                169 |                    3.3

benchmarks: to_json, data: FullBlock runs: 100, ms/run: 50, commit_hash: eea7e0764-dirty
mode       | µs/iteration | stdev µs/iteration % | avg iterations/run | stdev iterations/run %
to_json    | 219.71       |                 8.97 |                226 |                   5.42

compare: benchmark, old: 99275e6e3-dirty, new: eea7e0764-dirty
mode         | µs/iteration old | µs/iteration new | diff %      
to_json      | 827.98           | 294.22           | -64.46      

compare: full_block, old: 99275e6e3-dirty, new: eea7e0764-dirty
mode         | µs/iteration old | µs/iteration new | diff %      
to_json      | 445.7            | 219.71           | -50.7    
```