@arvidn 2022-05-18
> @arvidn Not sure if we need another benchmark here. Have you seen benchmarks/streamable.py?

I have seen it. I assumed it was benchmarking `Streamable` serialization, not json serialization.